[#132] refactor: 뷰모델 만들어서 가격 범위, 가격 평균에 대한 텍스트 코드 수정
